graphql-alt: Epoch.coinDenyList
## Description

Fetch the state of `0x403` at the start of the epoch. This object can be
modified during the epoch, but reads in that epoch will not be affected
by writes in the same epoch, so it is always safe to treat all reads in
an epoch as coming from this version of the object.

## Test plan

New E2E tests:

```
sui$ cargo nextest run         \
  -p sui-indexer-alt-e2e-tests \
  -- graphql/epochs/coin_deny_list
```